ajustando alguns erros de sintaxe e tipo

+++ src/Matrix.java
@@ -41,7 +41,7 @@ public class Matrix {
         return this.matrixElements[i][j];
     }
 
-    public void set(int i, int j, float value){
+    public void set(int i, int j, double value){
         this.matrixElements[i][j] = value;
     }
